+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - android 36.1.30
 - ios 26.2.10197
 - build: enable central package management (directory.packages.props)
+- build: use official bindings instead of manual jar import
 - Microsoft.Extensions.* 10.0.3
 - test : add colors to console test runner
 - test : use coreclr instead of mono